backend/requirements.txt
# João Macarrão Backend - Requirements
# Python 3.13, Django 5.2

# Core Django
Django==5.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0

# Database
psycopg2-binary==2.9.10

# Environment Variables
python-dotenv==1.0.1

# CORS
django-cors-headers==4.6.0

# Optional: Storage backends (descomente conforme necessário)
# django-cloudinary-storage==0.3.0
Pillow==11.0.0
# django-storages==1.14.4
# boto3==1.35.80

# Acessibilidade
google-cloud-texttospeech==2.17.2
django-filter==24.3

# Payment processing
stripe==11.2.0
mercadopago==2.2.3
qrcode==7.4.2

# Development tools (opcional)
# django-debug-toolbar==4.4.6
# pytest==8.3.4
# pytest-django==4.9.0
# black==24.10.0
# flake8==7.1.1

